+++ Service/OrderService.php
@@ -50,6 +50,8 @@ use MultiSafepay\ConnectCore\Util\CaptureUtil;
 
 class OrderService
 {
+    public const INVOICE_CREATE_AFTER_PARAM_NAME = 'multisaepay_create_inovice_after';
+
     /**
      * @var OrderRepositoryInterface
      */
@@ -304,7 +306,7 @@ class OrderService
 
         if ($this->canChangePaymentMethod($transactionType, $gatewayCode, $order)) {
             if ($this->giftcardUtil->isFullGiftcardTransaction($transaction)) {
-                $transactionType = $this->giftcardUtil->getGiftcardGatewayCodeFromTransaction($transaction) ?:
+                $transactionType = $this->giftcardUtil->getGiftcardGatewayCodeFromTransaction($transaction) ? :
                     $transactionType;
             }
 
@@ -474,6 +476,46 @@ class OrderService
         if ($order->canInvoice()) {
             if (!$this->captureUtil->isCaptureManualTransaction($transaction)) {
                 $this->invoiceService->invoiceByAmount($order, $payment, $transaction, $order->getBaseTotalDue());
+            $isCreateOrderAutomatically = $this->config->isCreateOrderInvoiceAutomatically($order->getStoreId());
+            $captureAmount = $order->getBaseTotalDue();
+            $orderId = $order->getIncrementId();
+            $payment->setTransactionId($transaction['transaction_id'] ?? '')
+                ->setAdditionalInformation(
+                    [
+                        PaymentTransaction::RAW_DETAILS => (array)$payment->getAdditionalInformation(),
+                        self::INVOICE_CREATE_AFTER_PARAM_NAME => !$isCreateOrderAutomatically,
+                    ]
+                )->setShouldCloseParentTransaction(false)
+                ->setIsTransactionClosed(0)
+                ->setIsTransactionPending(false);
+
+            $this->createInvoice($isCreateOrderAutomatically, $payment, $captureAmount, $orderId);
+            $payment->setParentTransactionId($transaction['transaction_id'] ?? '');
+            $payment->setIsTransactionApproved(true);
+            $this->orderPaymentRepository->save($payment);
+            $this->logger->logInfoForOrder($orderId, 'Payment saved', Logger::DEBUG);
+            $paymentTransaction = $payment->addTransaction(
+                PaymentTransaction::TYPE_CAPTURE,
+                null,
+                true
+            );
+
+            if ($paymentTransaction !== null) {
+                $paymentTransaction->setParentTxnId($transaction['transaction_id'] ?? '');
+            }
+
+            $paymentTransaction->setIsClosed(1);
+            $this->transactionRepository->save($paymentTransaction);
+            $this->logger->logInfoForOrder($orderId, 'Transaction saved', Logger::DEBUG);
+
+            if (!$isCreateOrderAutomatically) {
+                $order->addCommentToStatusHistory(
+                    __(
+                        'Captured amount %1 by MultiSafepay. Transaction ID: "%2"',
+                        $order->getBaseCurrency()->formatTxt($captureAmount),
+                        $paymentTransaction->getTxnId()
+                    )
+                );
             }
 
             // Set order processing
@@ -482,11 +524,37 @@ class OrderService
             $order->setStatus($status);
             $this->orderRepository->save($order);
             $this->logger->logInfoForOrder(
-                $order->getIncrementId(),
+                $orderId,
                 'Order status has been changed to: ' . $status,
                 Logger::DEBUG
             );
         }
+    }
+
+    /**
+     * @param bool $isCreateOrderAutomatically
+     * @param OrderPaymentInterface $payment
+     * @param float $captureAmount
+     * @param string $orderId
+     */
+    private function createInvoice(
+        bool $isCreateOrderAutomatically,
+        OrderPaymentInterface $payment,
+        float $captureAmount,
+        string $orderId
+    ): void {
+        if ($isCreateOrderAutomatically) {
+            $payment->registerCaptureNotification($captureAmount, true);
+            $this->logger->logInfoForOrder($orderId, 'Invoice created', Logger::DEBUG);
+
+            return;
+        }
+
+        $this->logger->logInfoForOrder(
+            $orderId,
+            'Invoice creation process was skipped by selected setting.',
+            Logger::DEBUG
+        );
     }
 
     /**
@@ -496,7 +564,7 @@ class OrderService
      * @throws ClientExceptionInterface
      * @throws Exception
      */
-    private function addInvoicesDataToTransactionAndSendEmail(
+    public function addInvoicesDataToTransactionAndSendEmail(
         OrderInterface $order,
         OrderPaymentInterface $payment,
         TransactionManager $transactionManager