+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug where filling in certain wrong values for the Date of birth field for AfterPay and in3 caused an error
 - Fixed a bug where "Unable to unserialize value" errors were logged for every product.
 
+### Added
+- Added possibility to skip automatic invoice creation after MultiSafepay payment.
+
 ## [2.8.2] - 2021-07-20
 ### Fixed
 - Fixed a bug where the default store url was being used for payment links in backend orders.